pyproject.toml
[tool.poetry]
name = "ohtu-2022"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}, {include = "tests"}]

[tool.pytest.ini_options]
pythonpath = [
  "tests"
]

[tool.poetry.dependencies]
python = "^3.8.1"
invoke = "^1.7.3"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
robotframework = "^6.0.1"
robotframework-seleniumlibrary = "^6.0.0"
gunicorn = "^20.1.0"
pybtex = "^0.24.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.0"
pytest = "^7.2.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
